backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
openai
pillow
celery[redis]
requests
websocket-client
redis
